DDCYLS-8871 Adding AES Encryption with Legacy Encryption Handling
