+++ it/test/uk/gov/hmrc/traderservices/services/SessionStateServiceISpec.scala
@@ -19,7 +19,8 @@ package uk.gov.hmrc.traderservices.services
 import org.apache.pekko.actor.ActorSystem
 import com.typesafe.config.Config
 import org.mongodb.scala.{MongoClient, MongoDatabase}
-import play.api.libs.json._
+import play.api.libs.json.*
+import uk.gov.hmrc.crypto.{Decrypter, Encrypter, SymmetricCryptoFactory}
 import uk.gov.hmrc.traderservices.support.AppISpec
 import uk.gov.hmrc.mongo.{CurrentTimestampSupport, MongoComponent}
 import uk.gov.hmrc.mongo.cache.{CacheItem, DataKey}
@@ -528,7 +529,6 @@ trait SessionStateServiceISpecSetup extends AppISpec {
     override val actorSystem: ActorSystem = app.injector.instanceOf[ActorSystem]
     override val cacheRepository = SessionStateServiceISpecSetup.this.cacheRepository
     lazy val keyProvider: KeyProvider = KeyProvider(app.injector.instanceOf[Config])
-    override val keyProviderFromContext: String => KeyProvider = _ => keyProvider
     override val stateFormats: Format[Int] = SessionStateServiceISpecSetup.this.stateFormats
     override def getJourneyId(journeyId: String): Option[String] = Option(journeyId)
     override val default: Int = 0
@@ -551,6 +551,8 @@ trait SessionStateServiceISpecSetup extends AppISpec {
     def getState(implicit rc: String): Future[Option[Int]] = currentSessionState.map(_.map(_._1))
     def setState(state: Int)(implicit rc: String): Future[Int] = super.save((state, Nil)).map(_._1)
 
+    override val legacyKeyProvider: KeyProvider = keyProvider
+    override val crypto: Encrypter & Decrypter = SymmetricCryptoFactory.aesGcmCrypto("UrI5kMAs7ewjByGBXD2+5+v3GZdCzutjTe07g37xc2M=")
   }
 
   trait TestService {
